New Builder API function, returns DasStream
The v2.3 builder function returned a list of datasets.  With the introduction
of vectors as first class objects with vector frame definitions in the stream
header, returning only the datasets is not enough information.  After all
packets are read and converted to DasDs a stream object containing only dataset
descriptors (not packet descriptors) is returned by the new API call:
  DasDsBldr_getStream()