Cubic dataset helpers and bug fix
This add new functions to make it easier to work with datasets
that have one coordinate physical dimension for each array
dimension, which is a common case.  The new functions are:

  DasDs_cubicCoords()
  DasDim_degenerate()

and a new iterator:

  struct das_cube_iter

In addition a type bug that was causing incorrect conversion
of TT2000 values to dastime for the milliseconds field has
been fixed.